refactored routeplanner
